+++ src/main/java/income_outgo/IncomeOutgoRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IncomeOutgoRepository extends JpaRepository<IncomeOutgo, Integer> {
+public interface IncomeOutgoRepository extends JpaRepository<IncomeOutgo, Long> {
 //    @Query("select * from income_outgo  where DATE_FORMAT(date,'%Y-%m') = :month")
 //    List<IncomeOutgo> findByMonth(@Param("month") String month);
-
 }
